examples/pytest-testing/pyproject.toml
[project]
name = "pytest-testing-example"
version = "0.1.0"
description = "Example demonstrating pytest integration with ni-datastore"
requires-python = ">=3.10,<4.0"
dependencies = [
    "pytest>=7.0.0",
    "pytest-check>=2.0.0",
    "python-dotenv>=1.0.0",
    "ni-datastore",
]

[tool.uv.sources]
ni-datastore = { path = "../..", editable = true }
